Remove RealisticSeasons integration
RealisticSeasons was formerly integrated with this plugin, and allowed for the current season to determine certain disasters. However, the plugin is paid, and the API is not separate from the plugin itself, so all usages of RealisticSeasons packages have been removed. Also, .gitignore has been updated to include the 'target' directory.